+++ doc/source/history.rst
@@ -8,6 +8,8 @@ Version history
 
 * **CHANGED:** a ``PerformanceParameters`` object needs to be specialized for the device used (by calling its ``for_device()`` method) before passing it to gates.
 
+* **CHANGED:** instead of using ``numpy.random.RandomState`` for key generation and encryption, ``DeterministicRNG`` and ``SecureRNG`` are available instead. The former is the wrapped ``RandomState``, fast, but not cryptographically secure; the latter is the secure random source provided by the OS, which can be rather slow.
+
 * ADDED: a high-level API hiding the Reikna details and removing some boilerplate.
 
 * FIXED: a bug in ``tlwe_noiseless_trivial()`` occasionally leading to memory corruption.